Move package to flight.
