+++ Compose-Flight-Search-App/app/src/main/java/com/example/flightsearchapp/ui/flight/FlightsListScreen.kt
@@ -1,10 +1,11 @@
-package com.example.flightsearchapp.ui
+package com.example.flightsearchapp.ui.flight
 
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import com.example.flightsearchapp.ui.ShowFlightUiState
 import com.example.flightsearchapp.ui.model.FlightModel
 import com.example.flightsearchapp.ui.theme.FlightSearchAppTheme
